Feat: Package Popup Dialog class
