+++ src/pages/scripts/packages/popup-dialog/README.md
@@ -1,0 +1,30 @@
+# Popup Dialog
+
+## Overview
+
+**Popup Dialog** is a JavaScript library for creating popup dialog elements.
+
+## Usage
+
+1. Import
+
+```js
+import { PopupDialog } from "@/popup-dialog/popup-dialog.js";
+```
+
+2. Use
+
+A sample code:
+```js
+const newDialog = new PopupDialog({
+    id: new Date(),
+    title: "New Dialog",
+    content: `<p>This is a new dialog.</p>`,
+});
+```
+
+## License
+
+Licensed under the [MIT License](./LICENSE).
+
+Copyright (C) よね/Yone
